Implement test to check connection to MySQL.
The objective of the feature is to build the MySQL environment.
To test the MySQL environment that was previously built, a test
of the connection to MySQL was implemented. This test also includes
checking the operation of various SQL operations.

Design doc: https://docs.google.com/document/d/1pRQqPizplE4UxKrOX-cU
tGXN13QGjPDUWhOfMF_eiXs/edit#heading=h.p1uhftpymtkl

Issue: #21
Test: sh test/test_mysql.sh